+++ src/query-builder/InsertQueryBuilder.ts
@@ -11,8 +11,7 @@ export class InsertQueryBuilder<Entity> extends QueryBuilder<Entity> {
     // -------------------------------------------------------------------------
 
     /**
-     * Specifies FROM which entity's table select/update/delete will be executed.
-     * Also sets a main string alias of the selection data.
+     * Specifies INTO which entity's table insertion will be executed.
      */
     into(entityTarget: Function|string): this {
         return this.setMainAlias(entityTarget);
@@ -34,18 +33,6 @@ export class InsertQueryBuilder<Entity> extends QueryBuilder<Entity> {
     values(values: ObjectLiteral|ObjectLiteral[]): this {
         this.expressionMap.valuesSet = values;
         return this;
-    }
-
-    /**
-     * Clones query builder as it is.
-     * Note: it uses new query runner, if you want query builder that uses exactly same query runner,
-     * you can create query builder this way: new InsertQueryBuilder(queryBuilder) where queryBuilder
-     * is cloned QueryBuilder.
-     */
-    clone(): InsertQueryBuilder<Entity> {
-        const qb = new InsertQueryBuilder<Entity>(this.connection);
-        qb.expressionMap = this.expressionMap.clone();
-        return qb;
     }
 
 }