+++ src/repository/Repository.ts
@@ -19,6 +19,8 @@ export class Repository<Entity extends ObjectLiteral> {
     // Protected Methods Set Dynamically
     // -------------------------------------------------------------------------
 
+    // todo: wny not to make them public?
+
     /**
      * Entity Manager used by this repository.
      */
@@ -203,6 +205,13 @@ export class Repository<Entity extends ObjectLiteral> {
      */
     async removeById(id: any, options?: RemoveOptions): Promise<void> {
         return this.manager.removeById(this.metadata.target, id, options);
+    }
+
+    /**
+     * Removes entity by a given entity id.
+     */
+    async removeByIds(ids: any[], options?: RemoveOptions): Promise<void> {
+        return this.manager.removeByIds(this.metadata.target, ids, options);
     }
 
     /**